requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF

# ----------------------
# Database
# ----------------------
flask-sqlalchemy
sqlalchemy-migrate

# ----------------------
# Celery
# ----------------------
celery

# ----------------------
# Production Server
# ----------------------
gunicorn
